Performance - Don't deepcopy Moves list.
At this point of time, I don't see the moves list ever changing for a given position (might change in future with different implementation). But for now, we don't need to copy the entire list. This gives almost a 15% improvement in Perft. The test was changed to better reflect the usecase - as if we GenerateMoves and restore, it doesn't make sense to go back to an empty list.