Performance - UnmakeMove() instead of saving all unnecessary states
SaveState() was a pretty heavy operation before this change - it was saving all Piece Boards, Occupancy boards etc. The problem was that there was no easy way to deepcopy these 2 dimensional arrays of reference types. Instead of copying all these unnecessary data, I instead only copy the Moves, which can be assigned via reference, and a few flags. The 2d arrays are computed with an UnmakeMove() function.